Migrated getTopics() from IxTheo to TuelibMixin and exhanced its functionality by also adding honourees.
